Update
updated - the registration activity no has a JSON object and simulates
an ok response from the server when user submits the registration.  Will
be adding the rest of my comments to this activity the next time I work
on it.

update- fixed an error within the username validation on the
registration activity with a nullpointer exception if you deleted all
text in the username field.

update- migrated all of my error message text into the strings file