updates
update - added libs folder for the DateandTime library im using as well
as the chart library.

update- added a weight tracking chart to the mystatus fragment.

update - the notifications menu item on the status screen should now
update the number next to it to reflect any user notifications.
Currently only tracking photo updates and weight updates that are
needed. isn't currently showing anything in the activity though just on
the menu.

update - fixed the time tracker for required updates and modified the
files and folders for photos to reflect the actual intention.

update - added the DateFormatsAndInfo class to handle formatting my
dates and parsing dates from strings for storage and comparison.