updates
update- added the review fragment activity to the menu.

update - the review activity is currently pulling in the photos that
have been flagged for future review. I don't like the hang up thats
occuring when I start the review activity I'm not sure how to fix it at
this point.

update - users can record unlimited weights but only one a day is saved.